backend/requirements.txt
Django>=5.1,<5.2
django-unfold>=0.40,<1.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0
django-filter>=24.0,<25.0
drf-spectacular>=0.27,<1.0
psycopg[binary]>=3.2,<4.0
celery>=5.4,<6.0
redis>=5.0,<6.0
django-celery-beat>=2.6,<3.0
Pillow>=10.4,<11.0
python-dotenv>=1.0,<2.0
gunicorn>=22.0,<23.0
WeasyPrint>=62.0,<63.0
# TCC Analytics Platform (Phase 1)
requests>=2.32,<3.0
lxml>=5.3,<6.0
